Add nlopt optimiser
