+++ quop_mpi/optimiser/__init__.py
@@ -1,3 +1,4 @@
-from .scipy_optimiser import ScipyOptimiser
+from .standard_optimisers import ScipyOptimiser
+from .standard_optimisers import NloptOptimiser
 
-__all__ = ["ScipyOptimiser"]
+__all__ = ["ScipyOptimiser", "NloptOptimiser"]
