--- quop_mpi/optimiser/scipy_optimiser.py
@@ -1,16 +1,0 @@
-from scipy.optimize import minimize
-
-
-class ScipyOptimiser():
-    def __init__(self, method="BFGS"):
-        self.objective=None
-        self.method=method
-
-        
-    def __call__(self, func, param, **kwargs):
-        self.objective = func
-        return(
-            minimize(self.objective,
-                     param,
-                     method=self.method))
-
